9.6 (1.1.13) added controller support

+++ src/main/java/me/antileaf/alice/cards/alice/Dessert.java
@@ -154,7 +154,7 @@ public class Dessert extends AbstractAliceCard {
 				float topOfCard = this.current_y + height;
 				float spacing = 50.0F * Settings.scale;
 				float centerY = topOfCard + spacing;
-				float sin = (float) Math.sin((double) (this.angle / 180.0F * Math.PI));
+				float sin = (float) Math.sin((this.angle / 180.0F * Math.PI));
 				float xOffset = sin * width;
 				sb.draw(
 						icon,
